requirements.txt
django

djangorestframework

psycopg2-binary

firebase-admin

langchain

langchain-google-genai

google-generativeai

python-dotenv

requests

google-cloud-vision

Pillow
